Refactor useEffects out into separate hooks
